Closes #2, det and linear equations with fractions
Determinant and Linear Equation Systems Solver can handle the use of
fractions and parentheses! The use of parentheses also involves the use
of multiplication `(2)(3) == 2(3) == (2)3 == 6`. There is a preference
in solving parentheses which is the innermost parentheses is solve
first. (eg. 2/(3/(4/5)); (4/5) is solved, then (3/(4/5)) and finally all
2/(3/(4/5))). It is not allowed to put a fraction after opening a
parentheses ('(/' is not allowed).