Add build phase script to retriev git info
The script retrieves the SHA, the current status (dirty or clean) and
the current branch. This is useful for developers, in the case where you
are working on a branch different than master, the branch name will be
shown in the about screen, plus a '*' to tipify a dirty status.