New Matrix Solver VIew Controller
Create a matrix of UITextFields in a UITableView in order to reduce memory
use in creating all the UITextFields (Specially large matrixes).
Depending on the chosen matrix solver it will display the correct
layout. (also I think it is optimized for iPhone 5, but that needs more
testing).

The use of a special UITableViewCell is needed, that is where
GIDAMatrixCell comes in, given an array it returns a cell with the as
many UITextFields as objects in the array, this array ideally contains
the place holder information. (Maybe this should be changed to only send
a number). If no new cell is created then just send an array to update
the placeholders and the text value of the UITextFields. Also the
information on who is the delegate for the UITextFields is required.

Deletes DeterminantSheetViewController, MatrixSheetViewController and
SelectionViewController as they are not needed any more.

This is related to #27