Layout modifications done in special method
There was a lot of code repeating, not it is in 1 method
`setViewsDimensionsFullScreen` and it only changes que dimensions if
needed (If going from a keyboard to keyboard does not recalculate and
change dimensions). Also set that when you want to start writing it
moves the row to the middle of the table for better view.